Change the polling to a post commit hook
To stop periodically polling the git repo from another thread, the
repository can update (call a git pull on it) itself by make a simple
get request to the wiki. To prevent this for abuse, a token have to be
defined in the raki.yml. This mechanism takes kare for multiple
namespaces/repositories. To call the url, use this syntax:

http://<host>/git.pull?namespace=<namespace>&token=<token>

Example: Put this in the post commit hook from the main repository to
update it on every commit:

curl -0 http://cthn.de/git.pull?namespace=default&token=somesecrettoken